+++ R/fractional.operators.R
@@ -214,7 +214,7 @@ fractional.operators <- function(L,
 #' domain of interest. Does not need to be given if either `mesh` or `graph` is supplied.
 #' @param C The mass matrix of a finite element discretization of the domain
 #' of interest. Does not need to be given if either `mesh` or `graph` is supplied.
-#' @param mesh An optional inla mesh. Replaces `d`, `C` and `G`.
+#' @param mesh An optional fmesher mesh. Replaces `d`, `C` and `G`.
 #' @param graph An optional `metric_graph` object. Replaces `d`, `C` and `G`. 
 #' @param range_mesh The range of the mesh. Will be used to provide starting values for the parameters. Will be used if `mesh` and `graph` are `NULL`, and if one of the parameters (kappa or tau for spde parameterization, or sigma or range for matern parameterization) are not provided.
 #' @param loc_mesh The mesh locations used to construct the matrices C and G. This option should be provided if one wants to use the `rspde_lme()` function and will not provide neither graph nor mesh. Only works for 1d data. Does not work for metric graphs. For metric graphs you should supply the graph using the `graph` argument. 
@@ -409,10 +409,19 @@ matern.operators <- function(kappa = NULL,
     stop("You should either provide mesh, graph, or provide both C *and* G!")
   }
 
+  if(!is.null(loc_mesh) && d!=1){
+    stop("loc_mesh only works with dimension 1.")
+  } else if (!is.null(loc_mesh)){
+    # mesh_1d <- fmesher::fm_mesh_1d(loc_mesh)
+    mesh_1d <- fm_mesh_1d(loc_mesh)
+  }
+
   if( (is.null(C) || is.null(G)) && (is.null(graph)) && (!is.null(loc_mesh) && d==1)){
-    fem <- rSPDE.fem1d(loc_mesh)
-    C <- fem$C
-    G <- fem$G
+    # fem <- rSPDE.fem1d(loc_mesh)
+    # fem <- fmesher::fm_fem(mesh_1d)
+    fem <- fm_fem(mesh_1d)
+    C <- fem$c0
+    G <- fem$g1
   }
 
   has_mesh <- FALSE
@@ -429,6 +438,28 @@ matern.operators <- function(kappa = NULL,
 
   if (!parameterization %in% c("matern", "spde")) {
     stop("parameterization should be either 'matern', 'spde' or 'graph'!")
+  }
+
+  if(parameterization == "spde"){
+    if(!is.null(nu)){
+      stop("For 'spde' parameterization, you should NOT supply 'nu'. You need to provide 'alpha'!")
+    }
+    if(!is.null(sigma)){
+      stop("For 'spde' parameterization, you should NOT supply 'sigma'. You need to provide 'tau'!")
+    }
+    if(!is.null(range)){
+      stop("For 'spde' parameterization, you should NOT supply 'range'. You need to provide 'kappa'!")
+    }
+  } else{
+    if(!is.null(alpha)){
+      stop("For 'matern' parameterization, you should NOT supply 'alpha'. You need to provide 'nu'!")
+    }
+    if(!is.null(tau)){
+      stop("For 'matern' parameterization, you should NOT supply 'tau'. You need to provide 'sigma'!")
+    }
+    if(!is.null(kappa)){
+      stop("For 'matern' parameterization, you should NOT supply 'kappa'. You need to provide 'range'!")
+    }
   }
 
   if(!is.null(graph)){
@@ -448,7 +479,9 @@ matern.operators <- function(kappa = NULL,
 
   if (!is.null(mesh)) {
     d <- get_inla_mesh_dimension(inla_mesh = mesh)
-    fem <- INLA::inla.mesh.fem(mesh)
+    # fem <- INLA::inla.mesh.fem(mesh)
+    # fem <- fmesher::fm_fem(mesh)
+    fem <- fm_fem(mesh)
     C <- fem$c0
     G <- fem$g1
     has_mesh <- TRUE
@@ -548,15 +581,19 @@ matern.operators <- function(kappa = NULL,
 
   if(!is.null(mesh)){
       make_A <- function(loc){
-        return(INLA::inla.spde.make.A(mesh = mesh, loc = loc))
+        # return(INLA::inla.spde.make.A(mesh = mesh, loc = loc))
+        # return(fmesher::fm_basis(x = mesh, loc = loc))
+        return(fm_basis(x = mesh, loc = loc))
       }
   } else if(!is.null(graph)){
       make_A <- function(loc){
-        return(graph$mesh_A(loc))
+        return(graph$fem_basis(loc))
       }
   } else if(!is.null(loc_mesh) && d == 1){
     make_A <- function(loc){
-          return(rSPDE::rSPDE.A1d(x = loc_mesh, loc = loc))
+          # return(rSPDE::rSPDE.A1d(x = loc_mesh, loc = loc))
+          # return(fmesher::fm_basis(x = mesh_1d, loc = loc))
+          return(fm_basis(x = mesh_1d, loc = loc))
           }   
   } else {
     make_A <- NULL
@@ -593,6 +630,20 @@ matern.operators <- function(kappa = NULL,
     output$range_mesh <- range_mesh
     output$graph <- graph
     output$loc_mesh <- loc_mesh
+    output$cov_function_mesh <- function(p, direct = FALSE){
+      if(is.null(output$make_A)){
+        stop("Please, create the object using the mesh.")
+      }
+      v <- output$make_A(loc = p)
+      if(direct){
+        return(output$Pr %*% solve(output$Q, output$Pr %*% t(v)))
+      } else{
+        return(Sigma.mult(output, t(v)))
+      } 
+    }
+    output$covariance_mesh <- function(){
+      return(output$Pr %*% solve(output$Q, output$Pr))
+  }    
     return(output)
   } else {
 
@@ -617,6 +668,22 @@ matern.operators <- function(kappa = NULL,
     out$range_mesh <- range_mesh
     out$graph <- graph
     out$loc_mesh <- loc_mesh    
+    out$cov_function_mesh <- function(p){
+      if(is.null(out$make_A)){
+        stop("Please, create the object using the mesh.")
+      }
+      v <- t(out$make_A(loc = p))
+      A <- Matrix::Diagonal(dim(C)[1])
+      v_bar <- kronecker(matrix(1, nrow = m + 1), v)
+      A_bar <- kronecker(matrix(1, ncol = m+1), A)
+      return((A_bar) %*% solve(out$Q, v_bar))
+    }
+
+  out$covariance_mesh <- function(){
+      A <- Matrix::Diagonal(dim(C)[1])
+      A_bar <- kronecker(matrix(1, ncol = m+1), A)
+      return((A_bar) %*% solve(out$Q, t(A_bar)))
+  }    
     return(out)
   }
 }
@@ -752,9 +819,13 @@ CBrSPDE.matern.operators <- function(C,
 
       if (d > 1) {
         if (compute_higher_order) {
-          fem <- INLA::inla.mesh.fem(mesh, order = m_alpha + 1)
+          # fem <- INLA::inla.mesh.fem(mesh, order = m_alpha + 1)
+          # fem <- fmesher::fm_fem(mesh, order = m_alpha + 1)
+          fem <- fm_fem(mesh, order = m_alpha + 1)
         } else {
-          fem <- INLA::inla.mesh.fem(mesh)
+          # fem <- INLA::inla.mesh.fem(mesh)
+          # fem <- fmesher::fm_fem(mesh)
+          fem <- fm_fem(mesh)
         }
 
         C <- fem$c0
@@ -765,10 +836,12 @@ CBrSPDE.matern.operators <- function(C,
         Gk <- list()
         Gk[[1]] <- G
         for (i in 2:m_order) {
-          Gk[[i]] <- fem[[paste0("g", i)]]
+            Gk[[i]] <- fem[[paste0("g", i)]]
         }
       } else if (d == 1) {
-        fem <- INLA::inla.mesh.fem(mesh, order = 2)
+        # fem <- INLA::inla.mesh.fem(mesh, order = 2)
+        # fem <- fmesher::fm_fem(mesh)
+        fem <- fm_fem(mesh)
         C <- fem$c0
         G <- fem$g1
         Ci <- Matrix::Diagonal(dim(C)[1], 1 / rowSums(C))
@@ -1047,10 +1120,10 @@ CBrSPDE.matern.operators <- function(C,
 #'
 #' # define a non-stationary range parameter
 #' kappa <- seq(from = 2, to = 20, length.out = length(x))
-#'
+#' alpha <- nu + 1/2
 #' # compute rational approximation
 #' op <- spde.matern.operators(
-#'   kappa = kappa, tau = tau, nu = nu,
+#'   kappa = kappa, tau = tau, alpha = alpha,
 #'   G = fem$G, C = fem$C, d = 1
 #' )
 #'
@@ -1095,6 +1168,28 @@ spde.matern.operators <- function(kappa = NULL,
     stop("parameterization should be either 'matern' or 'spde'!")
   }
 
+    if(parameterization == "spde"){
+    if(!is.null(nu)){
+      stop("For 'spde' parameterization, you should NOT supply 'nu'. You need to provide 'alpha'!")
+    }
+    if(!missing(B.sigma)){
+      stop("For 'spde' parameterization, you should NOT supply 'B.sigma'. You need to provide 'B.tau'!")
+    }
+    if(!missing(B.range)){
+      stop("For 'spde' parameterization, you should NOT supply 'B.range'. You need to provide 'B.kappa'!")
+    }
+  } else{
+    if(!is.null(alpha)){
+      stop("For 'matern' parameterization, you should NOT supply 'alpha'. You need to provide 'nu'!")
+    }
+    if(!missing(B.tau)){
+      stop("For 'matern' parameterization, you should NOT supply 'B.tau'. You need to provide 'B.sigma'!")
+    }
+    if(!missing(B.kappa)){
+      stop("For 'matern' parameterization, you should NOT supply 'B.kappa'. You need to provide 'B.range'!")
+    }
+  }
+
 if (is.null(d) && is.null(mesh) && is.null(graph)) {
     stop("You should give either the dimension d, the mesh or graph!")
   }
@@ -1103,10 +1198,18 @@ if (is.null(d) && is.null(mesh) && is.null(graph)) {
     stop("You should either provide mesh, graph, or provide both C *and* G!")
   }
 
+  if(!is.null(loc_mesh) && d!=1){
+    stop("loc_mesh only works with dimension 1.")
+  } else if (!is.null(loc_mesh)){
+    # mesh_1d <- fmesher::fm_mesh_1d(loc_mesh)
+    mesh_1d <- fm_mesh_1d(loc_mesh)
+  }
+
   if( (is.null(C) || is.null(G)) && (is.null(graph)) && (!is.null(loc_mesh) && d==1)){
-    fem <- rSPDE.fem1d(loc_mesh)
-    C <- fem$C
-    G <- fem$G
+    # fem <- rSPDE.fem1d(loc_mesh)
+    fem <- fm_fem(mesh_1d)
+    C <- fem$c0
+    G <- fem$g1
   }
 
   has_mesh <- FALSE
@@ -1121,7 +1224,9 @@ if (is.null(d) && is.null(mesh) && is.null(graph)) {
 
   if (!is.null(mesh)) {
       d <- get_inla_mesh_dimension(inla_mesh = mesh)
-      fem <- INLA::inla.mesh.fem(mesh)
+      # fem <- INLA::inla.mesh.fem(mesh)
+      # fem <- fmesher::fm_fem(mesh)
+      fem <- fm_fem(mesh)
       C <- fem$c0
       G <- fem$g1
       has_mesh <- TRUE
@@ -1287,11 +1392,13 @@ if (is.null(d) && is.null(mesh) && is.null(graph)) {
 
   if(!is.null(mesh)){
       make_A <- function(loc){
-        return(INLA::inla.spde.make.A(mesh = mesh, loc = loc))
+        # return(INLA::inla.spde.make.A(mesh = mesh, loc = loc))
+        # return(fmesher::fm_basis(x = mesh, loc=loc))
+        return(fm_basis(x = mesh, loc=loc))
       }
   } else if(!is.null(graph)){
       make_A <- function(loc){
-        return(graph$mesh_A(loc))
+        return(graph$fem_basis(loc))
       }
   } else if(!is.null(loc_mesh) && d == 1){
     make_A <- function(loc){
@@ -1338,6 +1445,20 @@ if (is.null(d) && is.null(mesh) && is.null(graph)) {
       output$B.sigma <- B.sigma
       output$B.range <- B.range
       output$parameterization <- parameterization
+    output$cov_function_mesh <- function(p, direct = FALSE){
+      if(is.null(output$make_A)){
+        stop("Please, create the object using the mesh.")
+      }
+      v <- output$make_A(loc = p)
+      if(direct){
+        return(output$Pr %*% solve(output$Q, output$Pr %*% t(v)))
+      } else{
+        return(Sigma.mult(output, t(v)))
+      } 
+    }
+    output$covariance_mesh <- function(){
+      return(output$Pr %*% solve(output$Q, output$Pr))
+  }    
   } else{
     type_rational_approximation <- type_rational_approximation[[1]]
     m_alpha <- floor(alpha)
@@ -1443,7 +1564,23 @@ if (is.null(d) && is.null(mesh) && is.null(graph)) {
   output$B.sigma <- B.sigma
   output$B.range <- B.range
   output$parameterization <- parameterization  
+  output$cov_function_mesh <- function(p){
+      if(is.null(output$make_A)){
+        stop("Please, create the object using the mesh.")
+      }
+      v <- t(output$make_A(loc = p))
+      A <- Matrix::Diagonal(dim(C)[1])
+      v_bar <- kronecker(matrix(1, nrow = m + 1), v)
+      A_bar <- kronecker(matrix(1, ncol = m+1), A)
+      return((A_bar) %*% solve(output$Q, v_bar))
+    }
   class(output) <- "CBrSPDEobj"
+  }
+
+  output$covariance_mesh <- function(){
+      A <- Matrix::Diagonal(dim(C)[1])
+      A_bar <- kronecker(matrix(1, ncol = m+1), A)
+      return((A_bar) %*% solve(output$Q, t(A_bar)))
   }
 
   return(output)